Run `rake compile` through bundler
Shouldn't make any difference, but I'm grasping at straws trying to
determine why Travis won't build this C extension. When building,
it claims the resulting librtlsdr.so is not recognized and possibly
truncated, which I cannot imagine being the case. Maybe rake-compiler
has a bug in the current version though?